requirements.txt
nanoid
pydantic
fastapi
uvicorn
pymongo
python-dotenv
pyjwt
dnspython